feat: выделил argparse и console в отдельную сущность
