+++ task.py
@@ -1,4 +1,7 @@
 class Task:
     def __init__(self, name):
+        """
+        :param name: текст таска
+        """
         self.name = name
         self.active = True
